feat: re-org blockchain model configs
- move blockchain org configs into the blockchain features.
- move general model config and types to core instead of lonely services
