+++ src/core/context/AppContextProvider.tsx
@@ -5,14 +5,12 @@ import { WalletStore } from '../../features/blockchain/stores/walletStore';
 import { TextStreamStore } from '../../core/stores/textStreamStore';
 import { ToolCallStreamStore } from '../../core/stores/toolCallStreamStore';
 import { MessageHistoryStore } from '../../core/stores/messageHistoryStore';
-import {
-  MODEL_REGISTRY,
-  SupportedBlockchainModels,
-  SupportedReasoningModels,
-} from '../../services/modelRegistry';
 import { ModelConfig } from './types';
-import { initializeMessageRegistry } from './initializeMessageRegistry';
+import { initializeMessageRegistry } from '../../features/blockchain/config/initializeMessageRegistry';
 import { useExecuteOperation } from './useExecuteOperation';
+import { MODEL_REGISTRY } from '../config/models';
+import { SupportedBlockchainModels } from '../../features/blockchain/config/models';
+import { SupportedReasoningModels } from '../../features/reasoning/config/models';
 
 export const AppContextProvider = ({
   children,
